Changes in preparation for Cliniface v6

+++ include/FaceTools/MultiFaceModelViewer.h
@@ -1,5 +1,5 @@
 /************************************************************************
- * Copyright (C) 2020 SIS Research Ltd & Richard Palmer
+ * Copyright (C) 2021 SIS Research Ltd & Richard Palmer
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -60,9 +60,9 @@ private:
     std::unordered_map<const FM*, QString> _attachedNames;
 
     void _updateModelLists( const FM*);
-    void _setViewerVisible( size_t, bool);
-    void _doOnViewerChanged( size_t, const Vis::FV*);
-    void _doOnComboBoxChanged( size_t, const QString&);
+    void _setViewerVisible( int, bool);
+    void _doOnViewerChanged( int, const Vis::FV*);
+    void _doOnComboBoxChanged( int, const QString&);
 
     MultiFaceModelViewer( const MultiFaceModelViewer&) = delete;
     void operator=( const MultiFaceModelViewer&) = delete;
